+++ pydoclint/utils/violation.py
@@ -21,33 +21,33 @@ VIOLATION_CODES = types.MappingProxyType({
         ' ).'
     ),
     104: (
-        'Arguments are the same in the docstring and the function signature,'
-        ' but are in a different order.'
+        'Arguments are the same in the docstring and the function'
+        ' signature, but are in a different order.'
     ),
-    105: 'Argument names match, but type hints in these args do not match:',
+    105: ('Argument names match, but type hints in these args do not match:'),
     106: (
-        'The option `--arg-type-hints-in-signature` is `True` but there are'
-        ' no argument type hints in the signature'
+        'The option `--arg-type-hints-in-signature` is `True` but'
+        ' there are no argument type hints in the signature'
     ),
     107: (
         'The option `--arg-type-hints-in-signature` is `True` but not all'
         ' args in the signature have type hints'
     ),
     108: (
-        'The option `--arg-type-hints-in-signature` is `False` but there are'
-        ' argument type hints in the signature'
+        'The option `--arg-type-hints-in-signature` is `False` but'
+        ' there are argument type hints in the signature'
     ),
     109: (
-        'The option `--arg-type-hints-in-docstring` is `True` but there are'
-        ' no type hints in the docstring arg list'
+        'The option `--arg-type-hints-in-docstring` is `True` but'
+        ' there are no type hints in the docstring arg list'
     ),
     110: (
         'The option `--arg-type-hints-in-docstring` is `True` but not all'
         ' args in the docstring arg list have type hints'
     ),
     111: (
-        'The option `--arg-type-hints-in-docstring` is `False` but there are'
-        ' type hints in the docstring arg list'
+        'The option `--arg-type-hints-in-docstring` is `False` but'
+        ' there are type hints in the docstring arg list'
     ),
     201: 'does not have a return section in docstring',
     202: (
@@ -59,20 +59,20 @@ VIOLATION_CODES = types.MappingProxyType({
         ' annotation.'
     ),
     301: (
-        '__init__() should not have a docstring; please combine it with the'
-        ' docstring of the class'
+        '__init__() should not have a docstring; please combine it'
+        ' with the docstring of the class'
     ),
     302: (
         'The class docstring does not need a "Returns" section, because'
         ' __init__() cannot return anything'
     ),
     303: (
-        'The __init__() docstring does not need a "Returns" section, because'
-        ' it cannot return anything'
+        'The __init__() docstring does not need a "Returns" section,'
+        ' because it cannot return anything'
     ),
     304: (
-        'Class docstring has an argument/parameter section; please put it in'
-        ' the __init__() docstring'
+        'Class docstring has an argument/parameter section; please'
+        ' put it in the __init__() docstring'
     ),
     305: (
         'Class docstring has a "Raises" section; please put it in the'
@@ -83,13 +83,13 @@ VIOLATION_CODES = types.MappingProxyType({
         ' __init__() cannot yield anything'
     ),
     307: (
-        'The __init__() docstring does not need a "Yields" section, because'
-        ' __init__() cannot yield anything'
+        'The __init__() docstring does not need a "Yields" section,'
+        ' because __init__() cannot yield anything'
     ),
     401: '',  # Deprecated
     402: (
-        'has "yield" statements, but the docstring does not have a "Yields"'
-        ' section'
+        'has "yield" statements, but the docstring does not have a'
+        ' "Yields" section'
     ),
     403: (
         'has a "Yields" section in the docstring, but there are no "yield"'
@@ -102,49 +102,58 @@ VIOLATION_CODES = types.MappingProxyType({
     ),
     405: (
         'has both "return" and "yield" statements. Please use'
-        ' Generator[YieldType, SendType, ReturnType] as the return type'
-        ' annotation, and put your yield type in YieldType and return type'
-        ' in ReturnType. More details in'
+        ' Generator[YieldType, SendType, ReturnType] as the'
+        ' return type annotation, and put your yield type in'
+        ' YieldType and return type in ReturnType. More details in'
         ' https://jsh9.github.io/pydoclint/notes_generator_vs_iterator.html'
     ),
     501: (
-        'has raise statements, but the docstring does not have a "Raises"'
-        ' section'
+        'has raise statements, but the docstring does not have a'
+        ' "Raises" section'
     ),
     502: (
-        'has a "Raises" section in the docstring, but there are not "raise"'
-        ' statements in the body'
+        'has a "Raises" section in the docstring, but there are not'
+        ' "raise" statements in the body'
     ),
     503: (
         'exceptions in the "Raises" section in the docstring do not match'
         ' those in the function body.'
     ),
     504: (
-        'has assert statements, but the docstring does not have a "Raises"'
-        ' section. (Assert statements could raise "AssertError".)'
+        'has assert statements, but the docstring does not have a'
+        ' "Raises" section. (Assert statements could raise'
+        ' "AssertError".)'
     ),
     601: (
-        'Class docstring contains fewer class attributes than actual class'
-        ' attributes.'
+        'Class docstring contains fewer class attributes than actual'
+        ' class attributes.'
     ),
     602: (
-        'Class docstring contains more class attributes than in actual class'
-        ' attributes.'
+        'Class docstring contains more class attributes than in actual'
+        ' class attributes.'
     ),
     603: (
         'Class docstring attributes are different from actual class'
         ' attributes.'
         ' (Or could be other formatting issues: '
-        'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103'
-        ' ).'
+        'https://jsh9.github.io/pydoclint/violation_codes.html'
+        '#notes-on-doc103 ).'
     ),
     604: (
         'Attributes are the same in docstring and class def, but are in a'
         ' different order.'
     ),
     605: (
-        'Attribute names match, but type hints in these attributes do not'
-        ' match:'
+        'Attribute names match, but type hints in these'
+        ' attributes do not match:'
+    ),
+    606: (
+        'This attribute is documented inline, but it should be documented'
+        ' in the class docstring instead.'
+    ),
+    607: (
+        'The class docstring does not need an "Attributes" section,'
+        ' because the class attributes are documented inline.'
     ),
 })