+++ tests/test_main.py
@@ -2108,3 +2108,222 @@ def testArgDefaults(
         list(map(str, violations))
         == expectedViolationsLookup[checkArgDefaults]
     )
+
+
+@pytest.mark.parametrize(
+    ('style', 'requireInlineClassVarDocs', 'argTypeHintsInDocstring'),
+    list(
+        itertools.product(
+            ['google', 'numpy', 'sphinx'], [False, True], [False, True]
+        )
+    ),
+)
+def testInlineClassAttributeDocs(
+        style: str,
+        requireInlineClassVarDocs: bool,
+        argTypeHintsInDocstring: bool,
+) -> None:
+    violations = _checkFile(
+        filename=DATA_DIR / style / 'class_attributes/inline_docstring.py',
+        style=style,
+        requireInlineClassVarDocs=requireInlineClassVarDocs,
+        argTypeHintsInDocstring=argTypeHintsInDocstring,
+        skipCheckingShortDocstrings=False,
+    )
+
+    # (requireInlineClassVarDocs, argTypeHintsInDocstring) -> expected violations
+    expectedViolationsLookup: dict[tuple[bool, bool], list[str]] = {
+        (False, False): [
+            'DOC606: Class `MyClass1`, Attribute `field1`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass1`, Attribute `field2`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass1`, Attribute `field3`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC601: Class `MyClass1`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass1`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not in the docstring: '
+            '[field1: int, field2: str, field3: list[int]]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC606: Class `MyClass2`, Attribute `documented_field`: This attribute is '
+            'documented inline, but it should be documented in the class docstring '
+            'instead.',
+            'DOC601: Class `MyClass2`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass2`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not in the docstring: '
+            '[documented_field: float, undocumented_field: bool]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC606: Class `MyClass3`, Attribute `field2`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass3`, Attribute `field3`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass3`, Attribute `field4`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass3`, Attribute `field5`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC601: Class `MyClass3`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass3`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not in the docstring: '
+            '[field2: str, field3: int, field4: bool]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC606: Class `MyClass4`, Attribute `field1`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC601: Class `MyClass4`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass4`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not in the docstring: '
+            '[field1: str]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+        ],
+        (False, True): [
+            'DOC606: Class `MyClass1`, Attribute `field1`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass1`, Attribute `field2`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass1`, Attribute `field3`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC601: Class `MyClass1`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass1`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not in the docstring: '
+            '[field1: int, field2: str, field3: list[int]]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC606: Class `MyClass2`, Attribute `documented_field`: This attribute is '
+            'documented inline, but it should be documented in the class docstring '
+            'instead.',
+            'DOC601: Class `MyClass2`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass2`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not in the docstring: '
+            '[documented_field: float, undocumented_field: bool]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC606: Class `MyClass3`, Attribute `field2`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass3`, Attribute `field3`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass3`, Attribute `field4`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC606: Class `MyClass3`, Attribute `field5`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC601: Class `MyClass3`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass3`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not in the docstring: '
+            '[field2: str, field3: int, field4: bool]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC606: Class `MyClass4`, Attribute `field1`: This attribute is documented '
+            'inline, but it should be documented in the class docstring instead.',
+            'DOC601: Class `MyClass4`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass4`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not in the docstring: '
+            '[field1: str]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+        ],
+        (True, False): [
+            'DOC601: Class `MyClass2`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass2`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not documented inline: '
+            '[undocumented_field: bool]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC607: Class `MyClass3`: The class docstring does not need an "Attributes" '
+            'section, because the class attributes are documented inline.',
+            'DOC601: Class `MyClass3`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass3`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not documented inline: '
+            '[field1: int]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+        ],
+        (True, True): [
+            'DOC601: Class `MyClass2`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass2`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not documented inline: '
+            '[undocumented_field: bool]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC607: Class `MyClass3`: The class docstring does not need an "Attributes" '
+            'section, because the class attributes are documented inline.',
+            'DOC601: Class `MyClass3`: Class docstring contains fewer class attributes '
+            'than actual class attributes.  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC603: Class `MyClass3`: Class docstring attributes are different from '
+            'actual class attributes. (Or could be other formatting issues: '
+            'https://jsh9.github.io/pydoclint/violation_codes.html#notes-on-doc103 ). '
+            'Attributes in the class definition but not documented inline: '
+            '[field1: int]. (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+            'DOC605: Class `MyClass4`: Attribute names match, but type hints in these '
+            'attributes do not match: field1  (Please read '
+            'https://jsh9.github.io/pydoclint/checking_class_attributes.html on how to '
+            'correctly document class attributes.)',
+        ],
+    }
+
+    assert (
+        list(map(str, violations))
+        == expectedViolationsLookup[
+            requireInlineClassVarDocs, argTypeHintsInDocstring
+        ]
+    )