+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
 
-## [0.8.4] - 2026-3-13
+## [0.8.4] - 2026-05-16
 
 - Added
+  - Support inline class-attribute docstrings, add `DOC606` and `DOC607`, and
+    introduce `--require-inline-class-var-docs` in native and flake8 modes
   - A new config option `--skip-checking-private-functions` (shortform:
     `-scpf`) so docstrings in private methods can be skipped
 - Full diff